# talde1erronka2/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/kategoriak": {
            "get": {
                "tags": [
                    "Kategoriak"
                ],
                "description": "Kategoria guztiak bueltatzen ditu.",
                "operationId": "1cb3c6ea847381bce167d94fb131f013",
                "responses": {
                    "200": {
                        "description": "Kategoria guztiak bueltatu ditu."
                    },
                    "404": {
                        "description": "Ez du kategoriarik topatu."
                    }
                }
            },
            "put": {
                "tags": [
                    "Kategoriak"
                ],
                "description": "Adierazitako informazioaren arabera kategoria bat eguneratzen du datubasean.",
                "operationId": "d6d613a7b44a5ce82a6f00abf770cfa9",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Kategoriaren ID-a",
                                        "type": "integer"
                                    },
                                    "izena": {
                                        "description": "Kategoriaren izena",
                                        "type": "string"
                                    },
                                    "eguneratze_data": {
                                        "description": "Kategoriaren eguneratze-data eta ordua",
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Kategoria ondo eguneratu da datubasean."
                    },
                    "404": {
                        "description": "Ez du kategoriarik topatu ID horrekin."
                    }
                }
            },
            "post": {
                "tags": [
                    "Kategoriak"
                ],
                "description": "Adierazitako informazioarekin kategoria berri bat txertatzen du datubasean.",
                "operationId": "1485e42964ea5aed827f93f4b180ab3e",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "izena": {
                                        "description": "Kategoriaren izena",
                                        "type": "string"
                                    },
                                    "sortze_data": {
                                        "description": "Kategoriaren sortze-data eta ordua",
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Kategoria ondo txertatu da datubasean."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Kategoriak"
                ],
                "description": "Adierazitako ID-a erabilita datubasean kategoriaren ezabatze logikoa egiten du.",
                "operationId": "2b895f8cb46077b64ea3e0e5e3c5cadc",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Kategoriaren ID-a",
                                        "type": "integer"
                                    },
                                    "ezabatze_data": {
                                        "description": "Kategoriaren ezabatze-data eta ordua",
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kategoriak ondo ezabatu dira datubasean."
                    },
                    "404": {
                        "description": "Ez dago kategoriarik ID horrekin."
                    }
                }
            }
        },
        "/kategoriak/{id}": {
            "get": {
                "tags": [
                    "Kategoriak"
                ],
                "description": "Adierazitako ID-aren arabera kategoria bat bueltatzen du.",
                "operationId": "25c2ef3d29d9388d6abe553745ceccb2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Kategoriaren ID-a",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aukeratutako kategoria bueltatzen du."
                    },
                    "404": {
                        "description": "Ez du kategoriarik topatu ID horrekin."
                    }
                }
            }
        },
...
    },
...
}